requirements.txt
PyYAML~=6.0.1
paramiko~=3.4.0
